+++ packages/did_plc/lib/src/entities/create_operation_v1.freezed.dart
@@ -12,7 +12,7 @@ part of 'create_operation_v1.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 CreateOperationV1 _$CreateOperationV1FromJson(Map<String, dynamic> json) {
   return _CreateOperationV1.fromJson(json);